fixed stupid call on void target error
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@71038 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/Eiffel/interface/graphical/tools/watch/es_watch_tool_panel.e
@@ -395,8 +395,10 @@ feature -- Memory management
 			recycle_expressions
 			watches_grid.reset_layout_manager
 			clean_watched_grid
-			update_commands_on_expressions_delayer.destroy
-			update_commands_on_expressions_delayer := Void
+			if update_commands_on_expressions_delayer /= Void then
+				update_commands_on_expressions_delayer.destroy
+				update_commands_on_expressions_delayer := Void
+			end
 		end
 
 feature {NONE} -- Memory management
